Fixing the error of the translations
Hi, I'm Ruah and you can view my profile at https://utopian.team I'm handling the contribution of translation category for open source project. I see a lot of mistakes on the current translation and started to fixed it, I can assure you that this translations of mine are the most accurate.